scheduler/requirements.txt
Flask-APScheduler==1.12.2
SQLAlchemy==1.4.22
sqlalchemy-stubs
gunicorn
psycopg2-binary
httpx
pydantic
pydantic[email]
orjson
aio-pika